refactor - moved all src code to 'src' dir (except main.py - need to run from main project dir)
